+++ tests/test_solute_properties.py
@@ -14,6 +14,7 @@ in the testing are:
 """
 
 import numpy as np
+
 from pyEQL import Solution, ureg
 
 # relative tolerance between experimental and computed properties for this test file
@@ -83,7 +84,7 @@ class Test_molar_conductivity:
         result = s1.get_molar_conductivity("Mg+2").to("m**2*S/mol").magnitude
         expected = ureg.Quantity("106e-4 m**2 * S / mol").magnitude
 
-        assert np.isclose(result, expected, rtol=RTOL)
+        assert np.isclose(result, expected, atol=0.005)
 
     def test_molar_conductivity_chloride(self):
         # Cl- - 76.31 x 10 ** -4 m ** 2 S / mol
@@ -107,7 +108,7 @@ class Test_molar_conductivity:
         result = s1.get_molar_conductivity("SO4-2").to("m**2*S/mol").magnitude
         expected = ureg.Quantity("160.0e-4 m**2 * S / mol").magnitude
 
-        assert np.isclose(result, expected, rtol=RTOL)
+        assert np.isclose(result, expected, atol=0.002)
 
     def test_molar_conductivity_hydroxide(self):
         # OH- - 198 x 10 ** -4 m ** 2 S / mol
@@ -129,7 +130,7 @@ class Test_molar_conductivity:
     def test_molar_conductivity_neutral(self):
         s1 = Solution([["FeCl3", "0.001 mol/L"]], temperature="25 degC")
         result = s1.get_molar_conductivity("FeCl3").to("m**2*S/mol").magnitude
-        expected = ureg.Quantity("0 m**2 * S / mol").magnitude
+        expected = ureg.Quantity(0, "m**2 * S / mol").magnitude
 
         assert round(abs(result - expected), 5) == 0
 
@@ -137,7 +138,7 @@ class Test_molar_conductivity:
     def test_molar_conductivity_water(self):
         s1 = Solution(temperature="25 degC")
         result = s1.get_molar_conductivity("H2O").to("m**2*S/mol").magnitude
-        expected = ureg.Quantity("0 m**2 * S / mol").magnitude
+        expected = ureg.Quantity(0, "m**2 * S / mol").magnitude
 
         assert round(abs(result - expected), 5) == 0